Tests for ioMmu using the right banks.

+++ src/remotes/mame/mameremote.ts
@@ -27,6 +27,9 @@ export class MameRemote extends DzrpQeuedRemote {
 	// The socket connection.
 	public socket: Socket;
 
+	// Set to true if disconnecting. Used to suppress errors.
+	public socketDisconnecting: boolean;
+
 	// Timeout between sending command and receiving response.
 	protected cmdRespTimeout?: NodeJS.Timeout;
 
@@ -54,6 +57,7 @@ export class MameRemote extends DzrpQeuedRemote {
 		// Init socket
 		this.socket = new Socket();
 		this.socket.unref();
+		this.socketDisconnecting = false;
 
 		// React on-open
 		this.socket.on('connect', async () => {
@@ -71,24 +75,30 @@ export class MameRemote extends DzrpQeuedRemote {
 
 		// Handle disconnect
 		this.socket.on('close', hadError => {
-			LogTransport.log('MameRemote: MAME terminated the connection: ' + hadError);
-			console.log('Close.');
-			// Error
-			const err = new Error('MameRemote: MAME terminated the connection!');
-			this.emit('error', err);
+			//console.log('Close.');
+			if (!this.socketDisconnecting) {
+				LogTransport.log('MameRemote: MAME terminated the connection: ' + hadError);
+				// Error
+				const err = new Error('MameRemote: MAME terminated the connection!');
+				this.emit('error', err);
+			}
 		});
 
 		// Handle errors
 		this.socket.on('error', err => {
-			LogTransport.log('MameRemote: Error: ' + err);
-			console.log('Error: ', err);
-			// Error
-			this.emit('error', err);
+			//console.log('Error: ', err);
+			if (!this.socketDisconnecting) {
+				LogTransport.log('MameRemote: Error: ' + err);
+				// Error
+				this.emit('error', err);
+			}
 		});
 
 		// Receive data
 		this.socket.on('data', data => {
-			this.dataReceived(data.toString());
+			if (!this.socketDisconnecting) {
+				this.dataReceived(data.toString());
+			}
 		});
 
 		// Start socket connection
@@ -143,17 +153,19 @@ export class MameRemote extends DzrpQeuedRemote {
 	 * Called e.g. when vscode sends a disconnectRequest
 	 */
 	public async disconnect(): Promise<void> {
+		if (!this.socket)
+			return;
+		this.socket.removeAllListeners();
+
 		// Send a k(ill) command
+		this.socketDisconnecting = true;
 		// TODO: Remove once MAME issue 9578 (https://github.com/mamedev/mame/issues/9578) 	is clarified:
-		//const resp =
-		await this.sendPacketData('k');
-		this.stopCmdRespTimeout();	// REMOVE with kill command
+		this.stopCmdRespTimeout();
+		await this.sendPacketData('k');	// REMOVE with kill command
 
 		return new Promise<void>(resolve => {
 			if (!this.socket)
 				return;
-
-			this.socket?.removeAllListeners();
 			// Timeout is required because socket.end() does not call the
 			// callback if it is already closed and the state cannot
 			// reliable be determined.
@@ -421,7 +433,7 @@ export class MameRemote extends DzrpQeuedRemote {
 
 			// Try to send immediately
 			if (this.messageQueue.length == 1)
-				this.sendNextMessage();
+				await this.sendNextMessage();
 		});
 	}
